HWC to CWH bigfix
Somehow when I was cleaning this up the other night I must have deleted a function and didn't rename some of the variables correctly (224 was supposed to be replaced with input_height, etc.). Finally the height and width somehow got swapped din the postprocess function - that's what I get for attempting to clean up code past midnight.  This commit fixes those minor bugs and replaces the Jetson image with the correct one. Also added a few more TODOs.